45: Added OrdersContainer component
