+++ src/containers/OrdersContainer/OrdersContainer.tsx
@@ -1,0 +1,56 @@
+import React, { FC, ReactElement } from 'react';
+
+import { CollectionTokenInfo, FullOrder, TokenInfo } from '@airswap/types';
+
+import NftCard from '../../components/NftCard/NftCard';
+import NftCardSkeleton from '../../components/NftCardSkeleton/NftCardSkeleton';
+import { getFullOrderReadableSenderAmountPlusTotalFees } from '../../entities/FullOrder/FullOrderHelpers';
+import { AppRoutes } from '../../routes';
+
+interface OrdersContainerProps {
+  currencyTokenInfo: TokenInfo;
+  orders: FullOrder[];
+  tokens: CollectionTokenInfo[];
+  className?: string;
+}
+
+const OrdersContainer: FC<OrdersContainerProps> = ({
+  currencyTokenInfo,
+  orders,
+  tokens,
+  className = '',
+}): ReactElement => (
+  <div className={`orders-container ${className}`}>
+    {orders
+      .map(order => {
+        const orderToken = tokens.find(token => token.id === +order.signer.id);
+        const price = getFullOrderReadableSenderAmountPlusTotalFees(order, currencyTokenInfo);
+
+        if (!orderToken) {
+          return (
+            <NftCardSkeleton
+              key={order.nonce}
+              price={price.toString()}
+              symbol={currencyTokenInfo.symbol}
+              to={`${AppRoutes.nftDetail}/${order.signer.id}`}
+              className="collection-widget__nft-card"
+            />
+          );
+        }
+
+        return (
+          <NftCard
+            key={order.nonce}
+            imageURI={orderToken.image}
+            price={price.toString()}
+            name={orderToken.name}
+            symbol={currencyTokenInfo.symbol}
+            to={`${AppRoutes.nftDetail}/${order.signer.id}`}
+            className="collection-widget__nft-card"
+          />
+        );
+      })}
+  </div>
+);
+
+export default OrdersContainer;
